Use LINES and COLS
We need to be careful about handling SIGWINCH.  curses will do
most of the work and return KEY_RESIZE, so we need to handle that.
This works pretty well as is, but the status bar is getting
left hanging mid-screen when the screen is sized larger.  Need
to think about what to do with words that are on screen when/if
a resize happens.  Or maybe just abort the game on a resize?
Maybe a resize should just invalidate the score (do we really
care about tracking high scores?)

type: ini