Read dictionary into memory at startup
Memory is cheap now, and getting the file into memory should not be
a problem.  However, this change is mostly put in place so that
we can get a more uniform distribution on words.  Currently,
there is a bias based on length of preceding word, and the word
that follows a 10-letter word is ~5 times more likely to be selected
that a word that follows a 2 letter word.  We need to do an
initial scan on the dictionary and find word boundaries and
then do a random selection on those indices.

type: ini