Put new words at the end of the list
The current mechanism use for deleting words is to
overwrite them with spaces.  By placing new words at
the end of the list, we will tend to replace words
from bottom to top, so that the spaces will not
overwrite words on the line immediately preceding.
This does not quite work with words of different
speeds, but is adequate for now.

type: ini