Get set_timer call out of main loop
The interaction with the timer is still very wonky.  I believe that
setting non-blocking (ie, timeout(-1)) is causing SIGALRM to
be handled differently than timeout(N) for large positive N,
but I do not fully understand it.  When we set timeout(-1),
the process_keys() call is made twice per interval, and words
drop twice as fast as they should.  Setting the interval larger
than the delay words, but setting it non-blocking does not.

type: ini

# Previously edited commit message:
# Wipo
# 
# type: wip
#