Remove invalid assertion
We now have routines that are clearing the word list, so
S->words can be NULL, and that is fine.  One long term goal
is to completely eliminate all of the dynamic memory
allocation that is taking place after initialization,
at which point we can clean up all of the data structures.

type: ini