Do not reset timer value
If the user repeatedly hits ctrl-n to increment the level, the timer
is getting reset to the new interval and the screen does not update
until the user stops hitting ctrl-n.  This change prevents that
behavior by not resetting the timer value when the interval is being
reset.

type: ux
type: ini