Reset timer in banner()
This change is mostly just cleaning up calls that are no
longer needed.  As a side-effect, the interval will remain
the same in bonus rounds rather than slowing down.  I had
thought it would be nice to start the bonus round slow and
increase with each word, and maybe that is true, but we
should clean up all the new_level() cruft before we make
that ux change.

type: ini