Call srandom
Most of the code invokes random(), but some parts have a pre-processor
macro that converts random to srand48.  At the moment, I am unsure how
seeding works between the different generators, so will just call
both.  There does not actually seem to be an issue, and this should
certainly be resolved and understood.

type: ini