Stop checking letters when finalizing
If we continue in the loop, then a key stroke will
potentially count twice.  It will be used to terminate
the word but will also be applied to the first character
of any word checked after the finalization.  In other
words, the loop in finalize() that clears out all of
the match values is invalidated.

type: ini
type: bug