Make decay_rate part of state
decay_rate is a value that should probably change dynamically,
but we will need to do some ux testing to see how exactly to
do that.  This value makes the game slightly faster than
the .95 in the previous commit, but still much slower at higher levels
that it had been prior to the previous commit.

type: ini
type: ux