Clear all keystrokes during banner
If the user hits a key during the display of "Game Over",
the high score display is immediately torn down.  There
is some interaction between SIGALRM and timeouts that
I do not understand; we should be able to remove all
timeouts and make all getch be fully blocking, but
if we do that the game move twice as fast.

type: ini